bootstrap.py
import logging
import mysql.connector
import handlers.users as user_handler
import handlers.passwords as password_handler
import handlers.auth as auth_handler
import asyncio
from aiogram import Bot
from aiogram.types import BotCommand
from aiogram.contrib.fsm_storage.memory import MemoryStorage
from aiogram import bot, dispatcher, executor
from aiogram.contrib.middlewares.logging import LoggingMiddleware
from aiogram import types
from aiogram.dispatcher.filters.state import StatesGroup, State
from aiogram import Dispatcher
from config import *
