requirements.txt
aiogram==2.16
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.7
coverage==6.1.2
frozenlist==1.2.0
furl==2.1.3
idna==3.3
jsonschema==4.2.1
multidict==5.2.0
multidimensional-urlencode==0.0.4
mysql-connector-python==8.0.27
nose==1.3.7
orderedmultidict==1.0.1
Pillow==8.4.0
pook==1.0.2
protobuf==3.19.1
pyrsistent==0.18.0
pytz==2021.3
pyzbar==0.1.8
requests==2.26.0
retailcrm==5.1.2
six==1.16.0
typing_extensions==4.0.0
urllib3==1.26.7
xmltodict==0.12.0
yarl==1.7.2


